refactor dna-contract and operation-contract
