--- blockchain/biobank-contract/1OrgLocalFabricOrg1GatewayConnection.json
@@ -1,43 +1,0 @@
-{
-    "certificateAuthorities": {
-        "org1ca-api.127-0-0-1.nip.io:8080": {
-            "url": "http://org1ca-api.127-0-0-1.nip.io:8080"
-        }
-    },
-    "client": {
-        "connection": {
-            "timeout": {
-                "orderer": "300",
-                "peer": {
-                    "endorser": "300"
-                }
-            }
-        },
-        "organization": "Org1"
-    },
-    "display_name": "Org1 Gateway",
-    "id": "org1gateway",
-    "name": "Org1 Gateway",
-    "organizations": {
-        "Org1": {
-            "certificateAuthorities": [
-                "org1ca-api.127-0-0-1.nip.io:8080"
-            ],
-            "mspid": "Org1MSP",
-            "peers": [
-                "org1peer-api.127-0-0-1.nip.io:8080"
-            ]
-        }
-    },
-    "peers": {
-        "org1peer-api.127-0-0-1.nip.io:8080": {
-            "grpcOptions": {
-                "grpc.default_authority": "org1peer-api.127-0-0-1.nip.io:8080",
-                "grpc.ssl_target_name_override": "org1peer-api.127-0-0-1.nip.io:8080"
-            },
-            "url": "grpc://org1peer-api.127-0-0-1.nip.io:8080"
-        }
-    },
-    "type": "gateway",
-    "version": "1.0"
-}
