+++ blockchain/biobank-contract/lib/dna-contract/dna-contract-contract.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const { ActiveContext, ActiveContract } = require('./../active-contract')
-const CryptoUtils = require('./../crypto-utils')
 const { v4: uuidv4 } = require('uuid');
 
 const DnaContract = require('./dna-contract.js');
@@ -9,9 +8,7 @@ const DnaContractList = require('./dna-contract-list.js');
 const OperationContract = require('./../operation/operation-contract')
 const OperationList = require('./../operation/operation-list')
 const DataList = require('./../data/data-list.js');
-const Data = require('../data/data.js');
-const Operation = require('../operation/operation');
-
+const DnaContractUtils = require('./dna-contract-utils.js')
 
 class DnaContractContext extends ActiveContext {
     constructor() {
@@ -28,8 +25,8 @@ class DnaContractContract extends ActiveContract {
     }
 
     async createDnaContract(ctx, dnaContractAttributes) {
-        const newDnaContractAttributes = handleDnaContractAttributes(dnaContractAttributes)
-        await validateContractCreation(ctx, newDnaContractAttributes)
+        const newDnaContractAttributes = DnaContractUtils.handleDnaContractAttributes(dnaContractAttributes)
+        await DnaContractUtils.validateContractCreation(ctx, newDnaContractAttributes)
         const dnaContract = DnaContract.createInstance(newDnaContractAttributes);
         await ctx.dnaContractList.addDnaContract(dnaContract);
         return dnaContract;
@@ -49,10 +46,8 @@ class DnaContractContract extends ActiveContract {
 
         if (operationType == 'buy_dna'){
             const dnaContract = await ctx.dnaContractList.getDnaContract(contractId);
-            let dna = await getData(ctx, dnaContract.dnaId)
-            // await BiocoinOperations.transferBiocoins(ctx, ctx.user.address, dna.collector, dnaContract.parameters.price)
-            // dna = await addOwnersInData(ctx, dna)
-            const operation = createBuyingOperation(ctx, dna, dnaContract)
+            let dna = await DnaContractUtils.getData(ctx, dnaContract.dnaId)
+            const operation = DnaContractUtils.createBuyingOperation(ctx, dna, dnaContract)
             return operation
         }
         return
@@ -64,84 +59,13 @@ class DnaContractContract extends ActiveContract {
         const args = [ "OperationPaymentContract:readOperationPayment",  operationId ]
         const payment = await this.queryCurrencyChannel(ctx, args)
         
-        checkPaymentAndOperation(ctx, payment, operation)
-        let dna = await getData(ctx, operation.details.data_id)
-        await addOwnersInData(ctx, dna)
+        DnaContractUtils.checkPaymentAndOperation(ctx, payment, operation)
+        let dna = await DnaContractUtils.getData(ctx, operation.details.data_id)
+        await DnaContractUtils.addOwnersInData(ctx, dna)
 
         return dna
     }
 }
 
-function handleDnaContractAttributes(dnaContractAttributes) {
-    const { dnaId, parameters, created_at } = JSON.parse(dnaContractAttributes);
-    const { price } = parameters
-    const filteredParameters = { price }
-    const id = CryptoUtils.getHash(dnaId)
-
-    const newDnaContractAttributes = {
-        id, dnaId, parameters: filteredParameters, created_at
-    }
-    return newDnaContractAttributes;
-}
-
-async function validateContractCreation(ctx, dnaContractAttributes){
-    const dna = await getData(ctx, dnaContractAttributes.dnaId)
-    if(dna.collector == undefined || ctx.user.address != dna.collector ){
-        throw new Error('Unauthorized')
-    }
-    // const dnaContractId = CryptoUtils.getHash(dna.id)
-    // const existingDnaContract = await ctx.dnaContractList.getDnaContract(dnaContractId);
-    // throw new Error(JSON.stringify(existingDnaContract))
-    // if(existingDnaContract != undefined){
-    //     throw new Error('The DNA contract already exists')
-    // }
-}
-
-async function addOwnersInData(ctx, dna){
-    if(dna.owners.includes(ctx.user.address) == false) {
-        dna.owners.push(ctx.user.address)
-        await ctx.dataList.updateState(dna);
-    }
-    return dna
-}
-
-async function getData(ctx, dataId){
-    const dataKey = Data.makeKey([dataId]);
-    const data = await ctx.dataList.getData(dataKey); 
-    return data
-}
-
-function checkPaymentAndOperation(ctx, payment, operation){
-    if(payment == undefined || payment.status != "paid"){
-        throw new Error("Operation not paid")
-    }
-    if (operation.userAddress != ctx.user.address){
-        throw new Error("user not allowed")
-    }
-    return true
-}
-
-async function createBuyingOperation(ctx, dna, dnaContract){
-    const operation = new OperationContract()
-    const id = uuidv4()
-    const operationAttributes = JSON.stringify({
-        type: 'buy',
-        userAddress: ctx.user.address,
-        created_at: new Date().toDateString(),
-        details: {
-            data_id: dna.id,
-            contractId: dnaContract.id
-        },
-        input: [{
-            address: ctx.user.address,
-            value: dnaContract.parameters.price
-        }],
-        output: [{
-            address: dna.collector,
-            value: dnaContract.parameters.price
-        }]
-    })
-    return await operation.createOperation(ctx, id, operationAttributes)
-}
 
 module.exports = DnaContractContract;
